Finished converting training to raster focal stats.
Not using terra::focal because that would calculate stats for every cell in the raster. Instead, using raster versions of the data and focal window, multiplying them, and getting the summary stats needed. That calculates the value for the central cell AERONET stations are in.
This should be consistent with the terra::focal approach used in prediction side (which needs to run for every cell in the AOI).